// src/types/demo.ts
export interface FormData {
  name: string;
  role: string;
  customRole: string;
  email: string;
  companyName: string;
  website: string;
  businessStage: string;
  fundingRange: string;
  employeeCount: string;
  revenue: string;
  industry: string;
  banks: string[];
  bookkeepingSystems: string[];
  customBookkeepingSystem: string;
  paymentGateways: string[];
  customPaymentGateway: string;
  demoDate: string;
  demoTime: string;
}